grpc/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
pika==1.3.2
protobuf==5.28.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==75.1.0
